codegen tests
