+++ tests/functional/syntax/modules/test_exports.py
@@ -309,61 +309,3 @@ exports: (lib1.bar, lib1.foo)
     assert e.value.prev_decl.col_offset == 9
     assert e.value.prev_decl.node_source_code == "lib1.foo"
     assert e.value.prev_decl.module_node.path == "main.vy"
-
-
-@pytest.fixture
-def simple_library(make_input_bundle):
-    iface = """
-@external
-def foo() -> uint256:
-    ...
-
-@external
-def bar() -> uint256:
-    ...
-    """
-    lib1 = """
-import ilib
-implements: ilib
-
-@external
-def foo() -> uint256:
-    return 1
-
-@external
-def bar() -> uint256:
-    return 2
-    """
-    return make_input_bundle({"lib1.vy": lib1, "ilib.vyi": iface})
-
-
-def test_exports_interface_simple(simple_library):
-    main = """
-import lib1
-
-exports: lib1.__interface__
-    """
-    out = compile_code(
-        main,
-        output_formats=["abi", "ast_dict"],
-        contract_path="main.vy",
-        input_bundle=simple_library,
-    )
-    fnames = [item["name"] for item in out["abi"]]
-    assert fnames == ["foo", "bar"]
-
-
-def test_exports_interface2(simple_library):
-    main = """
-import lib1
-
-exports: lib1.ilib
-    """
-    out = compile_code(
-        main,
-        output_formats=["abi", "ast_dict"],
-        contract_path="main.vy",
-        input_bundle=simple_library,
-    )
-    fnames = [item["name"] for item in out["abi"]]
-    assert fnames == ["foo", "bar"]
